refactor metaculus questions to use dist/scale
