+++ ergo/platforms/metaculus/question/linear.py
@@ -1,37 +1,24 @@
-import jax.numpy as np
+from dataclasses import dataclass
+from typing import Any, Dict
 
-from ergo.distributions import Logistic, LogisticMixture
+from ergo.distributions import Logistic, LogisticMixture, Scale
 
 from .continuous import ContinuousQuestion
 
 
+@dataclass
 class LinearQuestion(ContinuousQuestion):
     """
     A continuous Metaculus question that's on a linear (as opposed to a log) scale"
     """
 
-    def normalize_samples(self, samples):
-        """
-        Map samples from their true scale to the Metaculus normalized scale
+    scale: Scale
 
-        :param samples: samples from a distribution answering the prediction question
-            (true scale)
-        :return: samples on the normalized scale
-        """
-        return (samples - self.question_range["min"]) / (self.question_range_width)
-
-    def denormalize_samples(self, samples):
-        """
-        Map samples from the Metaculus normalized scale to the true scale
-
-        :param samples: samples on the normalized scale
-        :return: samples from a distribution answering the prediction question
-            (true scale)
-        """
-
-        # in case samples are in some other array-like format
-        samples = np.array(samples)
-        return self.question_range["min"] + (self.question_range_width) * samples
+    def __init__(
+        self, id: int, metaculus: Any, data: Dict, name=None,
+    ):
+        super().__init__(id, metaculus, data, name)
+        self.scale = Scale(self.question_range["min"], self.question_range["max"])
 
     # TODO: also return low and high on the true scale,
     # and use those somehow in logistic.py
@@ -43,11 +30,9 @@ class LinearQuestion(ContinuousQuestion):
         :param normalized_dist: normalized logistic distribution
         :return: logistic distribution on the true scale of the question
         """
-        scale_loc = (
-            normalized_dist.loc * self.question_range_width + self.question_range["min"]
-        )
+        scale_loc = normalized_dist.loc * self.scale.scale_range + self.scale.scale_min
 
-        true_scale = normalized_dist.scale * self.question_range_width
+        true_scale = normalized_dist.scale * self.scale.scale_range
         return Logistic(scale_loc, true_scale)
 
     def get_true_scale_mixture(
