+++ ergo/platforms/metaculus/question/log.py
@@ -1,65 +1,27 @@
-import math
+from dataclasses import dataclass
+from typing import Any, Dict
 
 from plotnine import scale_x_log10
+
+from ergo.distributions import LogScale
 
 from .continuous import ContinuousQuestion
 
 
+@dataclass
 class LogQuestion(ContinuousQuestion):
-    @property
-    def deriv_ratio(self) -> float:
-        return self.possibilities["scale"]["deriv_ratio"]
+
+    scale: LogScale
+
+    def __init__(
+        self, id: int, metaculus: Any, data: Dict, name=None,
+    ):
+        super().__init__(id, metaculus, data, name)
+        self.scale = LogScale(
+            self.question_range["min"],
+            self.question_range["max"],
+            self.possibilities["scale"]["deriv_ratio"],
+        )
 
     def _scale_x(self, xmin: float = None, xmax: float = None):
         return scale_x_log10(limits=(xmin, xmax))
-
-    def normalized_from_true_value(self, true_value) -> float:
-        """
-        Get a prediciton sample value on the normalized scale from a true-scale value
-
-        :param true_value: a sample value on the true scale
-        :return: a sample value on the normalized scale
-        """
-        shifted = true_value - self.question_range["min"]
-        numerator = shifted * (self.deriv_ratio - 1)
-        scaled = numerator / self.question_range_width
-        timber = 1 + scaled
-        floored_timber = max(timber, 1e-9)
-        return math.log(floored_timber, self.deriv_ratio)
-
-    def true_from_normalized_value(self, normalized_value):
-        """
-        Get a prediciton sample value on the true scale from a normalized-scale value
-
-        :param normalized_value: [description]
-        :type normalized_value: [type]
-        :return: [description]
-        :rtype: [type]
-        """
-        deriv_term = (self.deriv_ratio ** normalized_value - 1) / (self.deriv_ratio - 1)
-        scaled = self.question_range_width * deriv_term
-        return self.question_range["min"] + scaled
-
-    def normalize_samples(self, samples):
-        """
-        Map samples from the true scale to the normalized scale
-
-        :param samples: Samples on the true scale
-        :return: Samples on the normalized scale
-        """
-        if isinstance(samples, list):
-            return [self.normalized_from_true_value(sample) for sample in samples]
-        else:
-            return self.normalized_from_true_value(samples)
-
-    def denormalize_samples(self, samples):
-        """
-        Map samples from the normalized scale to the true scale
-
-        :param samples: Samples on the normalized scale
-        :return: Samples on the true scale
-        """
-        if isinstance(samples, list):
-            return [self.true_from_normalized_value(sample) for sample in samples]
-        else:
-            return self.true_from_normalized_value(samples)
